plumber
A plumber is responsible for the system of pipes within a building. This system transports water to faucets and appliances and also removes wastes. Plumbers also work with pipes that transport natural gas.

Plumbers install and maintain plumbing systems. These systems include storm water systems, sewer drainage, and fixture installation (such as sinks, bathtubs, and toilets). Plumbers must have great mechanical skills, be physically strong, and be comfortable working in small spaces. The plumbing trade also requires a firm understanding of basic math and geometry.

Plumbers must be certified in order to practice their trade. In the United States the training to earn a plumbing certificate is quite lengthy and varies from state to state. All states require an apprenticeship. An apprenticeship is when a student is taught a trade with hands-on experience by a master. Apprenticeships are highly competitive, so it is recommended that plumbing candidates first take courses at a trade school. Apprenticeships usually last about 4–5 years and also require hundreds of hours of classroom time.

After the apprenticeship, plumbers must pass an exam to become a journeyman plumber. Journeyman plumbers can work freely as a plumber. Plumbing is related to the trades of pipe fitting and steam fitting.

Canadian plumbers must also go through an apprenticeship. Certification is highly recommended even though only eight provinces require it. In addition to the separate provincial certification, Canada offers the Red Seal exam. The Red Seal exam allows plumbers to work anywhere in Canada without having to take additional tests.